feat: integrated clerk for login

+++ src/app/actions.ts
@@ -1,35 +1,47 @@
 "use server";
 
 import { prisma } from "@/utils/prisma";
-import { Prisma } from "@prisma/client";
+import { auth } from "@clerk/nextjs/server";
+import { MessageRole } from "@prisma/client";
 
 export async function getConversations() {
+  const { userId } = await auth();
+
   return await prisma.conversation.findMany({
+    where: { clerkId: userId },
     orderBy: { createdAt: "desc" },
   });
 }
 
 export async function getConversation(id: string) {
-  return await prisma.conversation.findUnique({ where: { id } });
+  const { userId } = await auth();
+
+  return await prisma.conversation.findUnique({
+    where: { id, clerkId: userId },
+  });
 }
 
 export async function createConversation(name: string) {
+  const { userId } = await auth();
+
   return await prisma.conversation.create({
-    data: { name },
+    data: { name, clerkId: userId },
   });
 }
 
 export async function updateConversation(id: string, name: string) {
+  const { userId } = await auth();
+
   return await prisma.conversation.update({
-    where: { id },
+    where: { id, clerkId: userId },
     data: { name },
   });
 }
 
 export async function deleteConversation(id: string) {
-  return await prisma.conversation.delete({
-    where: { id },
-  });
+  const { userId } = await auth();
+
+  return await prisma.conversation.delete({ where: { id, clerkId: userId } });
 }
 
 export async function getMessages(conversationId: string) {
@@ -47,7 +59,7 @@ export async function getMessagesNotInSummary(conversationId: string) {
 }
 
 export async function addMessagesWithConversationId(
-  messages: Prisma.MessageUncheckedCreateInput[]
+  messages: { role: MessageRole; content: string; conversationId: string }[]
 ) {
   return await prisma.message.createMany({ data: messages });
 }
@@ -56,15 +68,16 @@ export async function updateConversationSummary(
   conversationId: string,
   summary: string
 ) {
-  const conversation = await prisma.conversation.update({
+  return await prisma.conversation.update({
     where: { id: conversationId },
-    data: { summary },
+    data: {
+      summary,
+      messages: {
+        updateMany: {
+          where: { conversationId },
+          data: { isInSummary: true },
+        },
+      },
+    },
   });
-
-  await prisma.message.updateMany({
-    where: { conversationId },
-    data: { isInSummary: true },
-  });
-
-  return conversation;
 }
